helper: move ulog loading into helper with lru_cache
Since ULog parsing is an expensive operation, we now use one common method
in helper to parse ULog from wherever it's needed.
The method caches a configurable nr of logs in RAM, so that consecutive
readings of the same file are greatly speeded up. This also works between
different HTTP requests & sessions.