Handle -inf in PSD plots
In weird instances, if the raw values being used to create PSD
plots are exactly constant, the PSD will have 0's which underflow
when doing a log10 scale. This causes a failure when the JSON
is deserialized. We replace -inf with the smallest finite number
that's already in the PSD.