3d: update cesium to 1.123
In some cases it was possible to trigger a rendering exception
(when following the model and zooming in closely with small model size).

The separate Bing maps API key is not required anymore.
Also the orientation changed, the 90 deg yaw offset is not required anymore.
And something seems to have changed in the quaternion as well. Not quite
sure what exactly, but now it matches with the previous version.

Cesium changelog:
https://github.com/CesiumGS/cesium/blob/main/CHANGES.md?plain=1